finish adding gesture indicators
Gesture tutorial is added to the photo processing page. It includes
tap and swipe right gestures. The tap gesture appears first when the
image processing is complete, followed by the swipe right gesture. Once
the user has completed the tutorial, the user can tap as many times as
they want to listen to the audio before swiping right to send to friends.
Co-authored-by: Michael michaelleie0ca@gmail.com